#1043: split grpcapi/server_show.go (5,288 → 4,072 LOC) into per-RPC files
server_show.go held 16 functions: 14 RPC handlers/helpers + 2 file-scope
helpers (writeRPMConfig, firewallFilterTermExpansionCount) + the
3,957-LOC `ShowText` switch. This PR extracts the 12 non-`ShowText`
functions into per-RPC files. `ShowText` itself (which is 75% of the
original file) keeps its existing home; splitting it requires per-topic
case-by-case design and is left for a follow-up PR.

New files:
- server_show_status.go: GetStatus, GetGlobalStats, GetSystemInfo (262 LOC)
- server_show_zones.go: GetZones, GetPolicies, GetScreen (176 LOC)
- server_show_events.go: GetEvents (71 LOC)
- server_show_interfaces.go: GetInterfaces, ShowInterfacesDetail,
  showInterfacesTerse, writeKernelStats (720 LOC)
- server_show_forwarding.go: buildLocalForwarding, dialAndShowForwarding (45 LOC)

After split:
- server_show.go: 5,288 → 4,072 LOC (kept: helpers + ShowText)
- 5 new files, all under 1,000 LOC

`goimports -w` was used to prune the import set in each new file (each
inherited the full server_show.go import list and pruned unused ones).
The package-private helpers `writeRPMConfig` and
`firewallFilterTermExpansionCount` remain in server_show.go alongside
`ShowText`, which is their primary caller.

`go build ./...` clean. `go test ./pkg/grpcapi/...` passes.
